+++ chatapp/android/app/build.gradle
@@ -39,7 +39,8 @@ android {
     defaultConfig {
         // TODO: Specify your own unique Application ID (https://developer.android.com/studio/build/application-id.html).
         applicationId "work.tsumayojiproject.chatapp"
-        minSdkVersion 16
+        minSdkVersion 21
+        multiDexEnabled true
         targetSdkVersion 28
         versionCode flutterVersionCode.toInteger()
         versionName flutterVersionName
