+++ chatapp/pubspec.yaml
@@ -25,12 +25,12 @@ dependencies:
     sdk: flutter
 
   shared_preferences: any # UserDefaluts
-  firebase_core: ^0.2.5  # add dependency for firebase Core
-
-
-  # The following adds the Cupertino Icons font to your application.
-  # Use with the CupertinoIcons class for iOS style icons.
+  firebase_core:   ^0.2.5  # add dependency for firebase Core
+  cloud_firestore: ^0.8.2
   cupertino_icons: ^0.1.3
+
+flutter_localizations:
+  sdk: flutter
 
 dev_dependencies:
   flutter_test:
